IN-234 refactor: add BasicPost and PagesInfo
