+++ src/common/components/swiper/CustomSwiper.tsx
@@ -1,4 +1,5 @@
-import Image, { StaticImageData } from 'next/image'
+import { ImageUrl } from '@/common'
+import Image from 'next/image'
 import {
   EffectFade,
   HashNavigation,
@@ -9,17 +10,8 @@ import { Swiper, SwiperSlide } from 'swiper/react'
 
 import 'swiper/swiper-bundle.css'
 
-type ImageType = {
-  createdAt?: string
-  fileSize?: number
-  height?: number
-  uploadId?: string
-  url: string
-  width?: number
-}
-
 type ImagesTypes = {
-  images: Array<ImageType>
+  images: Array<{ url: ImageUrl }>
 }
 const CustomSwiper = ({ images }: ImagesTypes) => {
   if (images.length === 0) {
@@ -49,7 +41,7 @@ const CustomSwiper = ({ images }: ImagesTypes) => {
       {images.map((image, index) => (
         <SwiperSlide
           className={'w-full'}
-          key={image.url}
+          key={String(image.url)}
         >
           <Image
             alt={'picture from post'}
