+++ src/common/types.ts
@@ -19,10 +19,43 @@ export type PhotoResult = {
   photoUrl: Nullable<string>
 }
 
-export type PhotoFile = {
+export type ImageUrl = StaticImageData | string
+
+export type ImageFile = {
   createdAt: string
   fileSize: number
   height: number
-  url: StaticImageData | string
+  url: ImageUrl
   width: number
+}
+
+export type UploadImage = {
+  uploadId: string
+} & ImageFile
+
+export type FullName = {
+  firstName: string
+  lastName: string
+}
+
+export type BasicPost = {
+  avatarOwner: string
+  createdAt: string
+  description: string
+  id: number
+  images: UploadImage[]
+  isLiked: boolean
+  likesCount: number
+  location: string
+  owner: FullName
+  ownerId: number
+  updatedAt: string
+  userName: string
+}
+
+export type PagesInfo = {
+  page: number
+  pageSize: number
+  pagesCount: number
+  totalCount: number
 }
