+++ src/features/createPost/api/addPost.types.ts
@@ -1,12 +1,8 @@
-import { PhotoFile } from '@/common'
+import { BasicPost, UploadImage } from '@/common'
 
 export type UploadPhotoResponse = {
-  images: UploadPhoto[]
+  images: UploadImage[]
 }
-
-type UploadPhoto = {
-  uploadId: string
-} & PhotoFile
 
 export type UploadPost = {
   childrenMetadata: { uploadId: string }[]
@@ -14,20 +10,5 @@ export type UploadPost = {
 }
 
 export type UploadPostResponse = {
-  avatarOwner: string
   avatarWhoLikes: boolean
-  createdAt: string
-  description: string
-  id: number
-  images: UploadPhoto[]
-  isLiked: boolean
-  likesCount: number
-  location: string
-  owner: {
-    firstName: string
-    lastName: string
-  }
-  ownerId: number
-  updatedAt: string
-  userName: string
-}
+} & BasicPost
