+++ src/features/newsFeed/model/newsFeed.types.ts
@@ -1,40 +1,13 @@
+import { BasicPost } from '@/common'
+
 export type NewsFeedRoot = {
-  items: Post[]
+  items: BasicPost[]
   nextCursor: number
   page: number
   pageSize: number
   pagesCount: number
   prevCursor: number
   totalCount: number
-}
-
-export type Post = {
-  avatarOwner: string
-  createdAt: string
-  description: string
-  id: number
-  images: Image[]
-  isLiked: boolean
-  likesCount: number
-  location: string
-  owner: FullName
-  ownerId: number
-  updatedAt: string
-  userName: string
-}
-
-export type FullName = {
-  firstName: string
-  lastName: string
-}
-
-export type Image = {
-  createdAt: string
-  fileSize: number
-  height: number
-  uploadId: string
-  url: string
-  width: number
 }
 
 export type NewsFeedQuery = {
